feat: initializing first page
