+++ src/components/TextInput.jsx
@@ -1,27 +1,26 @@
 import { useState } from 'react';
 import styles from './TextInput.module.css'
 
-export default function TextInput({ placeholder = 'O que está acontecendo?', maxLength, ...props }) {
-    const [text, setText] = useState('');
+export default function TextInput(props) {
+    const [tweetList, setTweetList] = useState([]);
 
-    function onTextChange(event) {
-        const text = event.target.value;
-        if (text.length <= maxLength) {
-            setText(text)
-        }
-    }
+    // function sendTweet() {
+    //     setTweetList([...tweetList, text])
+    //     setText('');
+    // }
 
     return (
         <div>
-            <textarea 
-                className={styles.input} 
-                placeholder={placeholder} 
-                maxLength={maxLength} 
-                value={text} 
-                onChange={onTextChange} 
+            <textarea
+                className={styles.input}
                 {...props}
             />
-            <p>{text.length} / {maxLength}</p>
+            {/* <button onClick={sendTweet}>Enviar</button>
+            {tweetList.map((tweet) => { 
+                return (
+                    <p>{tweet}</p>
+                ) 
+            })} */}
         </div>
     )
 }
